requirements.txt
absl-py==2.3.1
altair==5.5.0
asttokens==2.4.1
astunparse==1.6.3
attrs==25.3.0
blinker==1.8.2
cachetools==6.1.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.13.1
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mysql-connector==2.2.9
mysql-connector-python==8.0.29
namex==0.1.0
narwhals==1.46.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==2.1.1
opt_einsum==3.4.0
optree==0.16.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
protobuf==5.29.5
psutil==6.0.0
pure_eval==0.2.3
pyArduino==0.2.2
pyarrow==20.0.0
pydeck==0.9.1
pyFirmata==1.1.0
Pygments==2.18.0
PyMySQL==1.0.2
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==306
pyzmq==26.2.0
referencing==0.36.2
regex==2024.9.11
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.46.1
streamlit-autorefresh==1.0.1
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.17.2
